requirements.txt
# template reqs
attrs==22.2.0
black==23.1.0
certifi
cfgv==3.3.1
click==8.1.3
distlib==0.3.6
exceptiongroup==1.1.0
filelock==3.9.0
identify==2.5.18
iniconfig==2.0.0
mypy==1.0.1
mypy-extensions==1.0.0
nodeenv==1.7.0
packaging==23.0
pathspec==0.11.0
platformdirs==3.0.0
pluggy==1.0.0
pre-commit==3.1.0
pytest==7.2.1
python-dotenv==0.21.1
PyYAML==6.0
ruff==0.0.252
tomli==2.0.1
typing_extensions==4.5.0
virtualenv==20.19.0

#Quest Gen specific reqs
transformers>=4.30.2
datasets>=2.12.0
evaluate>=0.4.0
numpy>=1.24.3
pandas>=2.0.1
torch>=2.0.1
nltk>=3.8.1
wandb>=0.15.3
accelerate>=0.20.1
absl-py>=1.4.0
rouge>=1.0.1
rouge-score>=0.1.2
bert_score>=0.3.13